opt: Convert exec code to use new opt expressions.
The exec.go file converts SQL AST nodes into the old opt expressions.
This PR converts that code to use the new opt expressions. As part
of this, I created a new `exec` package. That package will live in
the `opt` directory for now. But eventually, it will move to the
`sql` directory, and we will move other execution-related packages
into the `sql/exec` directory as well. Until then, it's better to
keep it here so that developers are not confused as to why we have a
top-level `exec` directory that doesn't actually contain the main
body of execution code.

Release note: None