Add implicit newline at end of actual output
Having to do this manually was an annoying papercut. The expectation
would always have a trailing newline except in the case in which it
ended the file (and had no trailing newline). This last case now
fails (and is rewritten into ending with a newline), which is deemed
acceptable.

This also fixes a bug in which a `-rewrite` of a file failing due to
a missing newline in the actual output would garble the file.