Ensure the test callback sees the sub-test instance
I found this API usage error while attempting to run t.Fatal or t.Skip
inside a test directive: it's not possible to change the status of an
outer test instance while `datadriven.RunTest` has instantiated a
sub-test.

Symptom:
```
 test executed panic(nil) or runtime.Goexit: subtest may have called FailNow on a parent test
```

This patch fixes it by giving the sub-test instance to the
callback. All usages must be updated.