Fix the generation of sub-test names
The DSL requires sub-test names to be prefixed by their
enclosing sub-test in the input, to clarify intent.

However the code was passing this string unmodified as first argument
to `t.Run`, which _also_ adds the parent name as prefix. The result
would be a doubly prefixed test name, for example `subtest hai/world`
would cause the sub-test to be named `hai/hai/world`, and
`hello/my/world` would become `hello/hello/my/hello/my/world`.

This patch fixes it.