feature: print out unified diff for failed tests
When there's a lot of output in a datadriven test, it can be challenging
to figure out what went wrong based on the output of the test if the
failure was just a subtle difference.

This commit introduces difflib and produces a unified diff when a test
fails that has more than 5 lines of output.