add TestData.MaybeScanArgs for optionally scanning a named argument
Add a new MaybeScanArgs to the TestData type to allow a caller to scan an
optional argument without manually searching for it within CmdArgs. This is a
common practice in the Pebble repo, and this interface extension will reduce
boilerplate.

Additionally add an Arg method that looks up the first instance of the named
argument, returning the CmdArg and bool indicating whether it was found.