fix 10s logging data race
The goroutine that logs every 10 seconds can race with the function
exiting, which can lead to a data race or even a panic because of
t.Log used after the test exits.

The fix is to block until the goroutine is exiting, by writing to the
unbuffered channel instead of closing it.